requirements.txt
Django==2.2
django-crispy-forms==1.12.0
gunicorn==20.1.0
Pillow==8.3.2
pytz==2021.1
sqlparse==0.4.2
